created a MovieRepo object and added two properties to the movie data class
