+++ Week4/BlockbusterRecycler/app/src/main/res/values/strings.xml
@@ -10,4 +10,14 @@
     <string name="summaryCaptain">Steve Rogers, a rejected military soldier, transforms into Captain America after taking a dose of a \"Super-Soldier serum\". But being Captain America comes at a price as he attempts to take down a war monger and a terrorist organization.</string>
     <string name="summaryWonder">When a pilot crashes and tells of conflict in the outside world, Diana, an Amazonian warrior in training, leaves home to fight a war, discovering her full powers and true destiny.</string>
     <string name="summaryEmpire">After the Rebels are brutally overpowered by the Empire on the ice planet Hoth, Luke Skywalker begins Jedi training with Yoda, while his friends are pursued by Darth Vader and a bounty hunter named Boba Fett all over the galaxy.</string>
+    <string name="starsTerminator">Arnold Schwarzenegger, Linda Hamilton, Michael Biehn</string>
+    <string name="starsGoT">Emilia Clarke, Peter Dinklage, Kit Harington</string>
+    <string name="starsStrangerThings">Millie Bobby Brown, Finn Wolfhard, Winona Ryder</string>
+    <string name="starsBackFuture">Michael J. Fox, Christopher Lloyd, Lea Thompson</string>
+    <string name="starsReturnKing">Elijah Wood, Viggo Mortensen, Ian McKellen</string>
+    <string name="starsEmpire">Mark Hamill, Harrison Ford, Carrie Fisher</string>
+    <string name="starsAvengers">Robert Downey Jr., Chris Evans, Scarlett Johansson</string>
+    <string name="starsCaptain">Chris Evans, Hugo Weaving, Samuel L. Jackson</string>
+    <string name="starsWonder">Gal Gadot, Chris Pine, Robin Wright</string>
+    <string name="starsIndy">Harrison Ford, Kate Capshaw, Ke Huy Quan</string>
 </resources>
